Add CLI support for extra repos in pbuilder.
This allows the user to create pbuilder environments with an additional
repository in the environment's sources.list file. To assist with
managing multiple such environments support has been added for providing
a name for the environment:

--extra-mirror "deb http://foo.bar trusty main"
--name foobar

To access these environments the --os-release and --name arguments
provided to pbuilder.shell/pbuilder.update must match those provided to
pbuilder.create.

Change-Id: I4ca2c520923f50a7a2e2f81796abac317bb0024c