Update patch set 1
Patch Set 1:

> (1 comment)

The instance will be cleaned up after it gets terminated as a part of the cleanup process at Melbourne. In retrospective, I should make it a flag in the function instead of always record the metadata, which would never be cleaned up if the instances are not terminated.

Patch-set: 1
Label: Verified=0