Support packaging based on just ubuntu version.
Currently packaging is based around only openstack version. This
is still the default but you can now pass in a ubuntu version as the
openstack release to build a non openstack version.

You can now also override the ubuntu release for the openstack release.
This will allow you to build eg. xenial-queens and bionic-queens packages.

Change-Id: If203e0dcf431746bb77ec54c77ab2a41b7076a15