Fix the missing trove instances issue
When searching the instances with project or user, the returned instances
list will not include trove instances, that is because trove instances are
globally launched by "trove" project. The real user's project id and user id
is under server.metadata field and nova list api can not handle this.

Fix this by searching the instance under "trove" project first.

Change-Id: Ic9a9c206cb927bb9f8ca9807c7b94b9848156d58