Add 'Member' role to bot_user account
This is needed to to create Magnum clusters.

It may be possible to update policy to allow for 'bot_user' role to do
such actions, but since the action of TenantManager adding a user to
their project means giving 'Member' role, having this role for a bot
should not be an issue.

Change-Id: Ieade0804bae4a58457c561bd23959fdaf2f67fd1